+++ src/main/scala/org/clulab/processors/bionlp/BioNLPTokenizerPostProcessor.scala
@@ -6,7 +6,6 @@ import edu.stanford.nlp.ling.CoreLabel
 import BioNLPTokenizerPostProcessor._
 import edu.stanford.nlp.process.CoreLabelTokenFactory
 
-import scala.StringBuilder
 import scala.collection.mutable
 import scala.collection.mutable.ArrayBuffer
 
@@ -15,7 +14,7 @@ import scala.collection.mutable.ArrayBuffer
  * User: mihais
  * Date: 11/16/14
  */
-class BioNLPTokenizerPostProcessor {
+class BioNLPTokenizerPostProcessor(val specialTokens:Set[String]) {
 
   def process(input:Array[CoreLabel]):Array[CoreLabel] = {
     var tokens = input
@@ -30,13 +29,14 @@ class BioNLPTokenizerPostProcessor {
     tokens = breakOnPattern(tokens, dashSuffixes)
 
     // break binary complexes
-    tokens = breakComplex(tokens, SINGLEDASH_PATTERN)
+    tokens = breakComplex(tokens, SINGLESLASHORDASH_PATTERN)
 
     // break mutations
     // TODO: this needs improvement, see Dane's comments
     tokens = breakMutant(tokens, SINGLEDASH_PATTERN)
 
     // break all (well, most) tokens containing a single slash; try to replace them with an enumeration
+    // does not apply to protein family names, which often contain slashes
     tokens = breakOneSlash(tokens, SINGLESLASH_PATTERN)
 
     if(CONTEXTUAL_COMPLEX_TOKENIZATION) {
@@ -56,6 +56,8 @@ class BioNLPTokenizerPostProcessor {
 
     tokens
   }
+
+  def isSpecialToken(s:String) = specialTokens.contains(s.toLowerCase)
 
   def breakOnPattern(tokens:Array[CoreLabel], pattern:Pattern):Array[CoreLabel] = {
     val output = new ArrayBuffer[CoreLabel]
@@ -92,14 +94,12 @@ class BioNLPTokenizerPostProcessor {
   def breakOneSlash(tokens:Array[CoreLabel], pattern:Pattern):Array[CoreLabel] = {
     val output = new ArrayBuffer[CoreLabel]
 
-//    print("Before breakOneSlash:")
-//    for(i <- tokens.indices) print(" " + tokens(i).word())
-//    println
-
     for(i <- tokens.indices) {
       val token = tokens(i)
       val matcher = pattern.matcher(token.word())
-      if (matcher.matches() && ! isMeasurementUnit(token.word())) {
+      if (matcher.matches() &&
+          ! isMeasurementUnit(token.word()) && // skip measurement units such as "ng/ml"
+          ! isSpecialToken(token.word())) { // skip special tokens such as family names containing slash such as "MEK1/MEK2"
         val sepPos = matcher.start(2)
         val s1 = token.word().substring(0, sepPos)
         val s3 = token.word().substring(sepPos + 1)
@@ -112,10 +112,6 @@ class BioNLPTokenizerPostProcessor {
         output += token
       }
     }
-
-//    print("After breakOneSlash:")
-//    for(i <- output.indices) print(" " + output(i).word())
-//    println
 
     output.toArray
   }
@@ -135,8 +131,8 @@ class BioNLPTokenizerPostProcessor {
       val token = tokens(i)
       val matcher = pattern.matcher(token.word())
       if (matcher.matches() && // contains a dash or some known separator
-        ((i < tokens.length - 1 && isComplex(tokens(i + 1).word())) || // followed by "complex", or
-          (i > 0 && isComplex(tokens(i - 1).word())))){ // preceded by "complex"
+        ((i < tokens.length - 1 && isComplex(tokens(i + 1).word())) || // is followed by "complex", or
+         (i > 0 && isComplex(tokens(i - 1).word())))){ // is preceded by "complex"
         val sepPos = matcher.start(2)
         val s1 = token.word().substring(0, sepPos)
         output += tokenFactory.makeToken(s1, token.beginPosition(), sepPos)
@@ -219,7 +215,7 @@ class BioNLPTokenizerPostProcessor {
       val matcher = pattern.matcher(token.word())
       if (matcher.matches() && // contains a dash or some known separator
         ((i < tokens.length - 1 && isMutant(tokens(i + 1).word())) || // followed by "mutant", or
-          (i > 0 && isMutant(tokens(i - 1).word())))){ // preceded by mutant
+         (i > 0 && isMutant(tokens(i - 1).word())))){ // preceded by mutant
         val sepPos = matcher.start(2)
         val s1 = token.word().substring(0, sepPos)
         output += tokenFactory.makeToken(s1, token.beginPosition(), sepPos)
@@ -249,7 +245,7 @@ class BioNLPTokenizerPostProcessor {
         crtToken.append(tokens(i + 1).word())
         i += 2
       } else {
-        if(crtToken.size > 0) {
+        if(crtToken.nonEmpty) {
           val word = crtToken.toString()
           output += tokenFactory.makeToken(word, crtTokenBeginPosition, word.length)
           crtToken = new StringBuilder
@@ -259,7 +255,7 @@ class BioNLPTokenizerPostProcessor {
         i += 1
       }
     }
-    if(crtToken.size > 0) {
+    if(crtToken.nonEmpty) {
       val word = crtToken.toString()
       output += tokenFactory.makeToken(word, crtTokenBeginPosition, word.length)
     }
@@ -328,14 +324,6 @@ class BioNLPTokenizerPostProcessor {
     output.toArray
   }
 
-  def isComplex(word:String):Boolean = {
-    val m = COMPLEX.matcher(word)
-    m.matches()
-  }
-  def isMutant(word:String):Boolean = {
-    val m = MUTANT.matcher(word)
-    m.matches()
-  }
 }
 
 object BioNLPTokenizerPostProcessor {
@@ -365,6 +353,7 @@ object BioNLPTokenizerPostProcessor {
 
   val SINGLESLASH_PATTERN = Pattern.compile(s"($VALID_PROTEIN)(/)($VALID_PROTEIN)", Pattern.CASE_INSENSITIVE)
   val SINGLEDASH_PATTERN = Pattern.compile(s"($VALID_PROTEIN_NO_DASH)(\\-)($VALID_PROTEIN_NO_DASH)", Pattern.CASE_INSENSITIVE)
+  val SINGLESLASHORDASH_PATTERN = Pattern.compile(s"($VALID_PROTEIN_NO_DASH)([\\-/])($VALID_PROTEIN_NO_DASH)", Pattern.CASE_INSENSITIVE)
 
   val SITE1 = Pattern.compile("[ACDEFGHIKLMNQRSTVWY]\\d+", Pattern.CASE_INSENSITIVE)
   val SITE2 = Pattern.compile("glycine|phenylalanine|leucine|serine|tyrosine|cysteine|tryptophan|proline|histidine|arginine|soleucine|methionine|threonine|asparagine|lysine|serine|arginine|valine|alanine|aspartate|glutamate|glycine", Pattern.CASE_INSENSITIVE)
@@ -384,6 +373,15 @@ object BioNLPTokenizerPostProcessor {
   def isParen(s:String) = PARENS.contains(s)
 
   def isMeasurementUnit(s:String):Boolean = MEASUREMENT_UNIT_WITHSLASH.matcher(s).matches()
+
+  def isComplex(word:String):Boolean = {
+    val m = COMPLEX.matcher(word)
+    m.matches()
+  }
+  def isMutant(word:String):Boolean = {
+    val m = MUTANT.matcher(word)
+    m.matches()
+  }
 
   def mkDashSuffixes:Pattern = {
     val allSuffixes = makeRegexOr(VALID_DASH_SUFFIXES)